requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
langchain-core>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20
openai>=1.3.0
tiktoken
requests
numpy
pydantic
regex
